+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 **Changed**:
 
 - **decidim-core**: Prevent many conversation with the same participants at the UI level. [\#2376](https://github.com/decidim/decidim/pull/2376)
+- **decidim-proposals**: Do not allow attachments on proposals edition [\#2221](https://github.com/decidim/decidim/pull/2221)
 
 **Fixed**: